Graph now displays a small range of the data, yet refreshes more frequently. Currently, the graph will crash after around 40 seconds. Will fix later. As well, accelerations and velocities are now saved as floats instead of doubles.

+++ app/src/main/java/com/example/d/culminatingproject/DataSet.java
@@ -1,6 +1,5 @@
 package com.example.d.culminatingproject;
 
-import android.hardware.Sensor;
 import android.os.Parcel;
 import android.os.Parcelable;
 
@@ -50,7 +49,7 @@ public class DataSet implements Comparable, Parcelable {
      * pre: none
      * post: new data point added to the set of points
      */
-    public void establishNextDataPoint(double xAcceleration, double yAcceleration, double zAcceleration) {
+    public void establishNextDataPoint(float xAcceleration, float yAcceleration, float zAcceleration) {
 
         if (!isFinished) {
             if(dataPoints.isEmpty()) {
